# update BenchmarkDotNet to latest version to take advantage of improvements