--- a/azure-pipelines.yml
+++ b/azure-pipelines.yml
@@ -188,7 +188,8 @@ jobs:
       csproj: src\benchmarks\micro\MicroBenchmarks.csproj
       runCategories: 'coreclr corefx'
       benchviewCategory: 'coreclr'
-      frameworks: # for private jobs we want to benchmark .NET Core 3.0 only
+      frameworks: # for private jobs we want to benchmark .NET Core 3.0 and 5.0 only
+        - netcoreapp5.0
         - netcoreapp3.0
       
   # Windows x86 micro benchmarks
@@ -203,7 +204,8 @@ jobs:
       csproj: src\benchmarks\micro\MicroBenchmarks.csproj
       runCategories: 'coreclr corefx'
       benchviewCategory: 'coreclr'
-      frameworks: # for private jobs we want to benchmark .NET Core 3.0 only
+      frameworks: # for private jobs we want to benchmark .NET Core 3.0 and 5.0 only
+        - netcoreapp5.0
         - netcoreapp3.0
 
   # Windows x64 ML.NET benchmarks
@@ -249,7 +251,8 @@ jobs:
       csproj: src/benchmarks/micro/MicroBenchmarks.csproj
       runCategories: 'coreclr corefx'
       benchviewCategory: 'coreclr'
-      frameworks: # for private jobs we want to benchmark .NET Core 3.0 only
+      frameworks: # for private jobs we want to benchmark .NET Core 3.0 and 5.0 only
+        - netcoreapp5.0
         - netcoreapp3.0
 
   # Ubuntu 1804 x64 ML.NET benchmarks