The update contains a lof of minor improvements and bug fixes, but the most important are:

### "Use new .NET Core 3.0 API to get the total number of allocated bytes for all threads" 

https://github.com/dotnet/BenchmarkDotNet/commit/f54055a23bbf8490499166ea90947c966619c26c So far BDN was using `GC.GetAllocatedBytesForCurrentThread` which returns the value for **current thread only**. Recently new API  was added, `GC.GetTotalAllocatedBytes` which returns the value of allocated managed memory for all threads. By using it we get the proper numbers. For multi-threaded benchmarks (few dozens in the enitre portfolio) we are going to get new, bigger numbers. This is a change of reported values, but also a **bug fix** (cc @stephentoub).

### "Use Median instead of Mean for overhead calculations" 

https://github.com/dotnet/BenchmarkDotNet/commit/d9901baacd89a212ee6d1e46fc6ec5dabc0e9d77  

One of the techniques BDN uses to allow for accurate nano-benchmarking is overhead deduction. BDN measures the overhead by benchmarking a method of same return type and arguments as the benchmarked method. The difference is that the "overhead" method does nothing for `void` and returns `default` for methods returning anything. 

So far the actual results was: 

```
Result(n) = Actual(n) - Average(Overhead)
```

But as @AndyAyersMS has spotted, using `Average` for deducing the overhead is bad when we have some outliers that affect the Average:

```log
1.69
1.65
1.70
1.83
2.60 <-- an outlier -->
1.76
1.69
1.69
1.73
1.64
1.66
1.67
1.71
1.68
1.64
```

The implementation was changed, now BDN uses Median.

How it's going to affect the reported results? The nano-benchmarks are going to get more stable results, we are going to have fewer results affected by outliers. 

### Power Plans

By default BDN is now enforcing High Performance Power plan when running benchmarks on Windows. https://github.com/dotnet/BenchmarkDotNet/commit/0dfa37efbba24f1384f2c9726f2dcce30827bc26

To make sure it does not affect the results reported in our lab, I have disabled this behaviour in the RecommendedConfig type (`.DontEnforcePowerPlan()`). So this new feature is not going to affect the results, we might consider enabling it in the futre.


### Other improvements

+ "make ArgumentsSource support IEnumerable<object[]> for benchmarks accepting a single argument to mimic MemberData behaviour" https://github.com/dotnet/BenchmarkDotNet/commit/ec296dc45de798f7407852d5ab7febe2b457eca4 to solve the issue @stephentoub has hit in #478
+ "make FullNameProvider public" https://github.com/dotnet/BenchmarkDotNet/commit/6d71308f4c1d96bcf8a526bfcc61bb23307b4041 (cc @billwert)


To see the diff for the update from  [0.11.3.1003](https://ci.appveyor.com/project/dotnetfoundation/benchmarkdotnet/builds/22341000) to  [0.11.5.1090](https://ci.appveyor.com/project/dotnetfoundation/benchmarkdotnet/builds/24750494) please click [here](https://github.com/dotnet/BenchmarkDotNet/compare/87d281d7dbf52036819efff52e6661e436648b73...6d71308f4c1d96bcf8a526bfcc61bb23307b4041)