@billwert 2019-05-27
Thanks @adamsitnik!

I have done a comparison between our current version of BDN and 0.11.5.1086, so very close to this diff.

I have put the total results here: https://gist.github.com/billwert/9469538fb651b8b022e7bebdf8c7295f

Of particular interest:

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality  |
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| ---------- |
| System.Memory.Constructors<Byte>.MemoryMarshalCreateSpan                         |     12.53 |             0.34 |             0.03 |           |
| System.Memory.Constructors_ValueTypesOnly<Int32>.ReadOnlyFromPointerLength       |     12.12 |             0.41 |             0.03 | multimodal|

These seem to be impacted by the shift from Mean to Median. Here's what the raw results look like for MemoryMarshalCreateSpan:
Candidate:

OverheadActual   1: 152565216 op, 242377100.00 ns, 1.5887 ns/op
OverheadActual   2: 152565216 op, 242047300.00 ns, 1.5865 ns/op
OverheadActual   3: 152565216 op, 244870100.00 ns, 1.6050 ns/op
OverheadActual   4: 152565216 op, 241318900.00 ns, 1.5817 ns/op
OverheadActual   5: 152565216 op, 243208200.00 ns, 1.5941 ns/op
OverheadActual   6: 152565216 op, 242559400.00 ns, 1.5899 ns/op
OverheadActual   7: 152565216 op, 247181200.00 ns, 1.6202 ns/op
OverheadActual   8: 152565216 op, 244912500.00 ns, 1.6053 ns/op
OverheadActual   9: 152565216 op, 242285700.00 ns, 1.5881 ns/op
OverheadActual  10: 152565216 op, 242764900.00 ns, 1.5912 ns/op
OverheadActual  11: 152565216 op, 246956300.00 ns, 1.6187 ns/op
OverheadActual  12: 152565216 op, 242917100.00 ns, 1.5922 ns/op
OverheadActual  13: 152565216 op, 242545600.00 ns, 1.5898 ns/op
OverheadActual  14: 152565216 op, 244401300.00 ns, 1.6019 ns/op
OverheadActual  15: 152565216 op, 245350300.00 ns, 1.6082 ns/op

WorkloadWarmup   1: 152565216 op, 243468100.00 ns, 1.5958 ns/op

// BeforeActualRun
WorkloadActual   1: 152565216 op, 242658700.00 ns, 1.5905 ns/op
WorkloadActual   2: 152565216 op, 242035800.00 ns, 1.5864 ns/op
WorkloadActual   3: 152565216 op, 245612600.00 ns, 1.6099 ns/op
WorkloadActual   4: 152565216 op, 242780700.00 ns, 1.5913 ns/op
WorkloadActual   5: 152565216 op, 244314200.00 ns, 1.6014 ns/op
WorkloadActual   6: 152565216 op, 242844700.00 ns, 1.5917 ns/op
WorkloadActual   7: 152565216 op, 242693400.00 ns, 1.5908 ns/op
WorkloadActual   8: 152565216 op, 242765900.00 ns, 1.5912 ns/op
WorkloadActual   9: 152565216 op, 241449100.00 ns, 1.5826 ns/op
WorkloadActual  10: 152565216 op, 241861600.00 ns, 1.5853 ns/op
WorkloadActual  11: 152565216 op, 247192900.00 ns, 1.6202 ns/op
WorkloadActual  12: 152565216 op, 248041500.00 ns, 1.6258 ns/op
WorkloadActual  13: 152565216 op, 247349100.00 ns, 1.6213 ns/op
WorkloadActual  14: 152565216 op, 244272200.00 ns, 1.6011 ns/op
WorkloadActual  15: 152565216 op, 242933700.00 ns, 1.5923 ns/op

// AfterActualRun
WorkloadResult   1: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult   2: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult   3: 152565216 op, 2695500.00 ns, 0.0177 ns/op
WorkloadResult   4: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult   5: 152565216 op, 1397100.00 ns, 0.0092 ns/op
WorkloadResult   6: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult   7: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult   8: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult   9: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult  10: 152565216 op, 0.00 ns, 0.0000 ns/op
WorkloadResult  11: 152565216 op, 4275800.00 ns, 0.0280 ns/op
WorkloadResult  12: 152565216 op, 5124400.00 ns, 0.0336 ns/op
WorkloadResult  13: 152565216 op, 4432000.00 ns, 0.0290 ns/op
WorkloadResult  14: 152565216 op, 1355100.00 ns, 0.0089 ns/op
WorkloadResult  15: 152565216 op, 16600.00 ns, 0.0001 ns/op

Baseline
OverheadActual   1: 124610960 op, 192520900.00 ns, 1.5450 ns/op
OverheadActual   2: 124610960 op, 255380400.00 ns, 2.0494 ns/op
OverheadActual   3: 124610960 op, 203979000.00 ns, 1.6369 ns/op
OverheadActual   4: 124610960 op, 193032100.00 ns, 1.5491 ns/op
OverheadActual   5: 124610960 op, 192476100.00 ns, 1.5446 ns/op
OverheadActual   6: 124610960 op, 193009900.00 ns, 1.5489 ns/op
OverheadActual   7: 124610960 op, 217588500.00 ns, 1.7461 ns/op
OverheadActual   8: 124610960 op, 216896600.00 ns, 1.7406 ns/op
OverheadActual   9: 124610960 op, 194498000.00 ns, 1.5608 ns/op
OverheadActual  10: 124610960 op, 202146800.00 ns, 1.6222 ns/op
OverheadActual  11: 124610960 op, 201542200.00 ns, 1.6174 ns/op
OverheadActual  12: 124610960 op, 185823800.00 ns, 1.4912 ns/op
OverheadActual  13: 124610960 op, 211242300.00 ns, 1.6952 ns/op
OverheadActual  14: 124610960 op, 194089800.00 ns, 1.5576 ns/op
OverheadActual  15: 124610960 op, 206162600.00 ns, 1.6544 ns/op
OverheadActual  16: 124610960 op, 223671300.00 ns, 1.7950 ns/op
OverheadActual  17: 124610960 op, 198295600.00 ns, 1.5913 ns/op
OverheadActual  18: 124610960 op, 213353900.00 ns, 1.7122 ns/op
OverheadActual  19: 124610960 op, 222495800.00 ns, 1.7855 ns/op
OverheadActual  20: 124610960 op, 216971000.00 ns, 1.7412 ns/op

WorkloadWarmup   1: 124610960 op, 262845500.00 ns, 2.1093 ns/op

// BeforeActualRun
WorkloadActual   1: 124610960 op, 247317800.00 ns, 1.9847 ns/op
WorkloadActual   2: 124610960 op, 265753300.00 ns, 2.1327 ns/op
WorkloadActual   3: 124610960 op, 252514900.00 ns, 2.0264 ns/op
WorkloadActual   4: 124610960 op, 268808100.00 ns, 2.1572 ns/op
WorkloadActual   5: 124610960 op, 271620900.00 ns, 2.1798 ns/op
WorkloadActual   6: 124610960 op, 262705200.00 ns, 2.1082 ns/op
WorkloadActual   7: 124610960 op, 240089500.00 ns, 1.9267 ns/op
WorkloadActual   8: 124610960 op, 261646300.00 ns, 2.0997 ns/op
WorkloadActual   9: 124610960 op, 274081100.00 ns, 2.1995 ns/op
WorkloadActual  10: 124610960 op, 238811500.00 ns, 1.9165 ns/op
WorkloadActual  11: 124610960 op, 227333800.00 ns, 1.8243 ns/op
WorkloadActual  12: 124610960 op, 257076200.00 ns, 2.0630 ns/op
WorkloadActual  13: 124610960 op, 263583000.00 ns, 2.1152 ns/op
WorkloadActual  14: 124610960 op, 238869900.00 ns, 1.9169 ns/op
WorkloadActual  15: 124610960 op, 242160900.00 ns, 1.9433 ns/op
WorkloadActual  16: 124610960 op, 254169800.00 ns, 2.0397 ns/op
WorkloadActual  17: 124610960 op, 238191600.00 ns, 1.9115 ns/op
WorkloadActual  18: 124610960 op, 225440600.00 ns, 1.8092 ns/op
WorkloadActual  19: 124610960 op, 242206900.00 ns, 1.9437 ns/op
WorkloadActual  20: 124610960 op, 252964200.00 ns, 2.0300 ns/op

// AfterActualRun
WorkloadResult   1: 124610960 op, 40558970.00 ns, 0.3255 ns/op
WorkloadResult   2: 124610960 op, 58994470.00 ns, 0.4734 ns/op
WorkloadResult   3: 124610960 op, 45756070.00 ns, 0.3672 ns/op
WorkloadResult   4: 124610960 op, 62049270.00 ns, 0.4979 ns/op
WorkloadResult   5: 124610960 op, 64862070.00 ns, 0.5205 ns/op
WorkloadResult   6: 124610960 op, 55946370.00 ns, 0.4490 ns/op
WorkloadResult   7: 124610960 op, 33330670.00 ns, 0.2675 ns/op
WorkloadResult   8: 124610960 op, 54887470.00 ns, 0.4405 ns/op
WorkloadResult   9: 124610960 op, 67322270.00 ns, 0.5403 ns/op
WorkloadResult  10: 124610960 op, 32052670.00 ns, 0.2572 ns/op
WorkloadResult  11: 124610960 op, 20574970.00 ns, 0.1651 ns/op
WorkloadResult  12: 124610960 op, 50317370.00 ns, 0.4038 ns/op
WorkloadResult  13: 124610960 op, 56824170.00 ns, 0.4560 ns/op
WorkloadResult  14: 124610960 op, 32111070.00 ns, 0.2577 ns/op
WorkloadResult  15: 124610960 op, 35402070.00 ns, 0.2841 ns/op
WorkloadResult  16: 124610960 op, 47410970.00 ns, 0.3805 ns/op
WorkloadResult  17: 124610960 op, 31432770.00 ns, 0.2522 ns/op
WorkloadResult  18: 124610960 op, 18681770.00 ns, 0.1499 ns/op
WorkloadResult  19: 124610960 op, 35448070.00 ns, 0.2845 ns/op
WorkloadResult  20: 124610960 op, 46205370.00 ns, 0.3708 ns/op

Note the actual results are often zero with thew new BDN. How should we handle this sort of test? Should we add inner iterations perhaps?