@adamsitnik 2019-05-27
I think that this particular update has not introduced a new problem, but instead revealed an existing one.

This update changes only the way we do the math, not the way we run the benchmarks.

I have put the "Candidate" numbers to excel to show the difference:

| Overhead(i) | Average     | Median | Workload(i) | Old Result   | New Result | New - Old   |
|-------------|-------------|--------|-------------|--------------|------------|-------------|
| 1,5887      | 1,597433333 | 1,5922 | 1,5905      | -0,006933333 | -0,0017    | 0,005233333 |
| 1,5865      |             |        | 1,5864      | -0,011033333 | -0,0058    | 0,005233333 |
| 1,605       |             |        | 1,6099      | 0,012466667  | 0,0177     | 0,005233333 |
| 1,5817      |             |        | 1,5913      | -0,006133333 | -0,0009    | 0,005233333 |
| 1,5941      |             |        | 1,6014      | 0,003966667  | 0,0092     | 0,005233333 |
| 1,5899      |             |        | 1,5917      | -0,005733333 | -0,0005    | 0,005233333 |
| 1,6202      |             |        | 1,5908      | -0,006633333 | -0,0014    | 0,005233333 |
| 1,6053      |             |        | 1,5912      | -0,006233333 | -0,001     | 0,005233333 |
| 1,5881      |             |        | 1,5826      | -0,014833333 | -0,0096    | 0,005233333 |
| 1,5912      |             |        | 1,5853      | -0,012133333 | -0,0069    | 0,005233333 |
| 1,6187      |             |        | 1,6202      | 0,022766667  | 0,028      | 0,005233333 |
| 1,5922      |             |        | 1,6258      | 0,028366667  | 0,0336     | 0,005233333 |
| 1,5898      |             |        | 1,6213      | 0,023866667  | 0,0291     | 0,005233333 |
| 1,6019      |             |        | 1,6011      | 0,003666667  | 0,0089     | 0,005233333 |
| 1,6082      |             |        | 1,5923      | -0,005133333 | 1E-04      | 0,005233333 |

As you can see, here there was no outliers and Average is almost same as Median. The difference is very small (0,005233333). The results are not affected.

The same for "Baseline" numbers:

| Overhead(i) | Average | Median  | Workload(i) | Old Result | New Result | New - Old |
|-------------|---------|---------|-------------|------------|------------|-----------|
| 1,545       | 1,65923 | 1,62955 | 1,9847      | 0,32547    | 0,35515    | 0,02968   |
| **2,0494**      |         |         | 2,1327      | 0,47347    | 0,50315    | 0,02968   |
| 1,6369      |         |         | 2,0264      | 0,36717    | 0,39685    | 0,02968   |
| 1,5491      |         |         | 2,1572      | 0,49797    | 0,52765    | 0,02968   |
| 1,5446      |         |         | 2,1798      | 0,52057    | 0,55025    | 0,02968   |
| 1,5489      |         |         | 2,1082      | 0,44897    | 0,47865    | 0,02968   |
| 1,7461      |         |         | 1,9267      | 0,26747    | 0,29715    | 0,02968   |
| 1,7406      |         |         | 2,0997      | 0,44047    | 0,47015    | 0,02968   |
| 1,5608      |         |         | 2,1995      | 0,54027    | 0,56995    | 0,02968   |
| 1,6222      |         |         | 1,9165      | 0,25727    | 0,28695    | 0,02968   |
| 1,6174      |         |         | 1,8243      | 0,16507    | 0,19475    | 0,02968   |
| 1,4912      |         |         | 2,063       | 0,40377    | 0,43345    | 0,02968   |
| 1,6952      |         |         | 2,1152      | 0,45597    | 0,48565    | 0,02968   |
| 1,5576      |         |         | 1,9169      | 0,25767    | 0,28735    | 0,02968   |
| 1,6544      |         |         | 1,9433      | 0,28407    | 0,31375    | 0,02968   |
| **1,795**       |         |         | 2,0397      | 0,38047    | 0,41015    | 0,02968   |
| 1,5913      |         |         | 1,9115      | 0,25227    | 0,28195    | 0,02968   |
| 1,7122      |         |         | 1,8092      | 0,14997    | 0,17965    | 0,02968   |
| **1,7855**      |         |         | 1,9437      | 0,28447    | 0,31415    | 0,02968   |
| 1,7412      |         |         | 2,03        | 0,37077    | 0,40045    | 0,02968   |

Here, we had some outliers (bold numbers) and the difference between Average and Median is noticable: 0,02968ns (it's still a little difference). This is the case @AndyAyersMS has reported and @AndreyAkinshin has fixed by using the Median instead of Average.

I will try to find some time this week and answer the most important question: **why do we get zeros**? Are the two mentioned benchmarks optimized by JIT to a const? Are they single instruction and BDN can't handle that  and we should rewrite them? Are they very dependent on the array alignment and we should rewrite them? Is BDN generating the right code for the overhead method that returns Span?